refactor(scraper): generalize custom scraper
